refactor: 불필요한 필드 제거
