+++ src/main/java/baseball/Application.java
@@ -1,23 +1,18 @@
 package baseball;
 
 import baseball.context.ApplicationContext;
+import baseball.controller.GameController;
 import baseball.domain.BaseBallGame;
-import baseball.ui.GameController;
 
 public class Application {
     public static void main(String[] args) {
         GameController controller = ApplicationContext.getGameController();
-
-        run(controller);
-    }
-
-    private static void run(GameController controller) {
         do {
-            doGame(controller);
+            doOneGame(controller);
         } while (controller.willRestartGame());
     }
 
-    private static void doGame(GameController controller) {
+    private static void doOneGame(GameController controller) {
         BaseBallGame game = controller.startGame();
         while (controller.isProceeding(game)) {
             controller.guess(game);
