+++ src/main/java/baseball/context/ApplicationContext.java
@@ -1,9 +1,8 @@
 package baseball.context;
 
-import baseball.domain.NumberGenerateStrategy;
+import baseball.controller.GameController;
 import baseball.domain.Opponent;
 import baseball.domain.RandomBallNumbersGenerator;
-import baseball.ui.GameController;
 import baseball.ui.InputView;
 import baseball.ui.OutputView;
 
@@ -13,31 +12,20 @@ public class ApplicationContext {
     private static final ApplicationContext applicationContext;
 
     static {
-        RandomBallNumbersGenerator numberGenerateStrategy = new RandomBallNumbersGenerator();
-        Opponent opponent = new Opponent(numberGenerateStrategy);
+        Opponent opponent = new Opponent(new RandomBallNumbersGenerator());
+        InputView inputView = new InputView(new Scanner(System.in));
+        OutputView outputView = new OutputView();
 
-        applicationContext = new ApplicationContext(
-                opponent,
-                numberGenerateStrategy,
-                new GameController(new InputView(new Scanner(System.in)), new OutputView(), opponent)
-        );
+        applicationContext = new ApplicationContext(new GameController(inputView, outputView, opponent));
     }
 
     public static GameController getGameController() {
         return applicationContext.gameController;
     }
 
-    private final Opponent opponent;
-    private final NumberGenerateStrategy numberGenerateStrategy;
     private final GameController gameController;
 
-    private ApplicationContext(
-            Opponent opponent,
-            NumberGenerateStrategy numberGenerateStrategy,
-            GameController gameController
-    ) {
-        this.opponent = opponent;
-        this.numberGenerateStrategy = numberGenerateStrategy;
+    private ApplicationContext(GameController gameController) {
         this.gameController = gameController;
     }
 }
